Generalize data-source deeplink and merge BEMANIWiki section
- Add djdx://update?type=dataSource&id=<id> to refresh any external data
  source (textage, sdvxIn, wikiIidx, bm2dx, wikiDdr); replaces the DEBUG
  reload-ddr-meta action
- Extract all reload logic into ExternalDataReloader, shared by the
  External Data Sources screen and the deeplink
- Merge the beatmania IIDX and DanceDanceRevolution BEMANIWiki items into a
  single BEMANIWiki 2nd section, named by game
